Relocated the password code into a different file.
Originally, I had placed the password file maintenance code in with the
general resource indexing code. This seemed ill-adviced. So relocated.
